# src/main/java/com/eaglebank/api/dto/error/ErrorResponse.java
package com.eaglebank.api.dto.error;

public class ErrorResponse {

  private String message;

  public ErrorResponse() {
  }

  public ErrorResponse(String message) {
    this.message = message;
  }

  public String getMessage() {
    return message;
  }

  public void setMessage(String message) {
    this.message = message;
  }
}